README.md
This is the testing readme for the COMPUTER SCIENCE exercise. 

Define two variables and assign them values of 100 and 29. After this, make the interpreter sum up the two numbers and multiply their result by 3. Calculate the 2nd exponent of the previous number and save it to a new variable. Make the program print the result in the following way: The result of the calculation was 149769 Make sure you use valid variable names and name your variables meaningfully. Make sure that output of your program has indentation as shown in the above example